[skip ci] :recycle: api: move mass import template route to src + add tests
