+++ api/lib/application/certification-centers/index.js
@@ -351,31 +351,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/certification-centers/{certificationCenterId}/import',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserIsMemberOfCertificationCenter,
-            assign: 'isMemberOfCertificationCenter',
-          },
-          {
-            method: securityPreHandlers.checkCertificationCenterIsNotScoManagingStudents,
-            assign: 'isCertificationCenterNotScoManagingStudents',
-          },
-        ],
-        validate: {
-          params: Joi.object({ certificationCenterId: identifiersType.certificationCenterId }),
-        },
-        handler: certificationCenterController.getSessionsImportTemplate,
-        tags: ['api', 'sessions'],
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Elle permet de récupérer le fichier de création de sessions de certification',
-        ],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/certification-centers/{certificationCenterId}/invitations',
       config: {
         pre: [
