+++ api/tests/certification/session/unit/application/session-mass-import-route_test.js
@@ -114,4 +114,23 @@ describe('Unit | Router | session-mass-import-route', function () {
       });
     });
   });
+
+  describe('GET /api/certification-centers/{certificationCenterId}/import', function () {
+    it('should exist', async function () {
+      // given
+      sinon.stub(securityPreHandlers, 'checkUserIsMemberOfCertificationCenter').callsFake((_, h) => h.response(true));
+      sinon
+        .stub(securityPreHandlers, 'checkCertificationCenterIsNotScoManagingStudents')
+        .callsFake((_, h) => h.response(true));
+      sinon.stub(sessionMassImportController, 'getSessionsImportTemplate').returns('ok');
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/certification-centers/123/import');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+  });
 });
