+++ api/tests/unit/application/certification-centers/index_test.js
@@ -268,23 +268,4 @@ describe('Unit | Router | certification-center-router', function () {
       sinon.assert.calledOnce(certificationCenterController.findPaginatedSessionSummaries);
     });
   });
-
-  describe('GET /api/certification-centers/{certificationCenterId}/import', function () {
-    it('should exist', async function () {
-      // given
-      sinon.stub(securityPreHandlers, 'checkUserIsMemberOfCertificationCenter').callsFake((_, h) => h.response(true));
-      sinon
-        .stub(securityPreHandlers, 'checkCertificationCenterIsNotScoManagingStudents')
-        .callsFake((_, h) => h.response(true));
-      sinon.stub(certificationCenterController, 'getSessionsImportTemplate').returns('ok');
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/certification-centers/123/import');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
 });
